refactor(#792): extract recent-projects helpers into project_ops_recents.rs (sweep)
project_ops.rs 618 -> 471 (leaves the debt list). The recent-projects list
management + project path/name helpers (sync_recent_projects, register/mark,
recent_project_items, canonical_project_path, project_display_name,
parse_path_argument) move to project_ops_recents.rs; the session/config/device
lifecycle stays. All pub(crate), re-exported from project_ops so crate::project_ops::*
and the persistence tests' super:: paths resolve unchanged. adapter-gui lib tests
(540) pass, zero warnings.
