+++ crates/adapter-gui/src/lib.rs
@@ -153,6 +153,7 @@ mod model_search;
 mod model_search_wiring;
 mod preset_search;
 mod project_ops;
+mod project_ops_recents;
 // #679: `pub` so the issue_599 integration test can reach
 // `block_type_picker_items`. A private mod made `cargo test --tests` (and thus
 // `cargo llvm-cov`) fail to compile, which silently zeroed all coverage.
