+++ crates/adapter-gui/src/project_ops.rs
@@ -2,17 +2,24 @@ use application::command::Command;
 use application::dispatcher::CommandDispatcher;
 
 use crate::state::{AppConfigYaml, ProjectPaths, ProjectSession};
-use crate::RecentProjectItem;
-use crate::{AppWindow, UNTITLED_PROJECT_NAME};
+use crate::AppWindow;
 use anyhow::Result;
 use domain::ids::DeviceId;
-use infra_filesystem::{AppConfig, FilesystemStorage, GuiAudioDeviceSettings, RecentProjectEntry};
+use infra_filesystem::{AppConfig, FilesystemStorage, GuiAudioDeviceSettings};
 use infra_yaml::{load_chain_preset_file, ChainBlocksPreset};
 use project::device::DeviceSettings;
 use project::project::Project;
 use std::env;
 use std::fs;
 use std::path::{Path, PathBuf};
+
+// Issue #792 split: recent-projects + path/name helpers live in
+// project_ops_recents.rs. Re-exported so crate::project_ops::* and the
+// super:: paths in the persistence test modules keep resolving.
+pub(crate) use crate::project_ops_recents::{
+    canonical_project_path, mark_recent_project_invalid, parse_path_argument, project_display_name,
+    recent_project_items, register_recent_project, sync_recent_projects,
+};
 
 pub(crate) fn open_cli_project(path: &PathBuf) -> Result<ProjectSession> {
     if !path.exists() {
@@ -44,144 +51,6 @@ pub(crate) fn load_and_sync_app_config() -> Result<AppConfig> {
         application::app_config_persist::persist_app_config_snapshot(config.clone());
     }
     Ok(config)
-}
-
-pub(crate) fn sync_recent_projects(config: &mut AppConfig) -> bool {
-    let original = config.clone();
-    let mut synced = Vec::new();
-    for recent in &config.recent_projects {
-        let path = PathBuf::from(&recent.project_path);
-        // Skip path.exists() check here — it can block indefinitely on
-        // disconnected network volumes or external drives (macOS stat hang).
-        // Validity is checked lazily when the user tries to open the project.
-        let canonical_path = if path.is_absolute() {
-            path.clone()
-        } else {
-            env::current_dir()
-                .map(|d| d.join(&path))
-                .unwrap_or(path.clone())
-        };
-        let canonical_path_string = canonical_path.to_string_lossy().to_string();
-        if synced
-            .iter()
-            .any(|current: &RecentProjectEntry| current.project_path == canonical_path_string)
-        {
-            continue;
-        }
-        synced.push(RecentProjectEntry {
-            project_path: canonical_path_string,
-            project_name: if recent.project_name.trim().is_empty() {
-                UNTITLED_PROJECT_NAME.to_string()
-            } else {
-                recent.project_name.clone()
-            },
-            is_valid: true,
-            invalid_reason: None,
-        });
-    }
-    config.recent_projects = synced;
-    *config != original
-}
-
-pub(crate) fn canonical_project_path(path: &PathBuf) -> Result<PathBuf> {
-    // Do NOT call path.exists() here — blocks on disconnected network volumes.
-    // fs::canonicalize resolves symlinks and normalises the path without blocking
-    // for paths that exist on local storage; for paths that don't exist it errors
-    // and we fall back to the raw path.
-    if let Ok(c) = fs::canonicalize(path) {
-        return Ok(c);
-    }
-    if path.is_absolute() {
-        return Ok(path.clone());
-    }
-    Ok(env::current_dir()?.join(path))
-}
-
-pub(crate) fn register_recent_project(config: &mut AppConfig, path: &PathBuf, name: &str) {
-    let canonical_path = canonical_project_path(path).unwrap_or(path.clone());
-    let path_string = canonical_path.to_string_lossy().to_string();
-    config
-        .recent_projects
-        .retain(|current| current.project_path != path_string);
-    config.recent_projects.insert(
-        0,
-        RecentProjectEntry {
-            project_path: path_string,
-            project_name: if name.trim().is_empty() {
-                UNTITLED_PROJECT_NAME.to_string()
-            } else {
-                name.trim().to_string()
-            },
-            is_valid: true,
-            invalid_reason: None,
-        },
-    );
-}
-
-pub(crate) fn mark_recent_project_invalid(config: &mut AppConfig, path: &PathBuf, reason: &str) {
-    let canonical_path = canonical_project_path(path).unwrap_or(path.clone());
-    let path_string = canonical_path.to_string_lossy().to_string();
-    if let Some(recent) = config
-        .recent_projects
-        .iter_mut()
-        .find(|current| current.project_path == path_string)
-    {
-        recent.is_valid = false;
-        recent.invalid_reason = Some(if reason.trim().is_empty() {
-            "Projeto inválido".to_string()
-        } else {
-            reason.trim().to_string()
-        });
-    }
-}
-
-pub(crate) fn recent_project_items(
-    recent_projects: &[RecentProjectEntry],
-    query: &str,
-) -> Vec<RecentProjectItem> {
-    let query = query.trim().to_lowercase();
-    recent_projects
-        .iter()
-        .enumerate()
-        .filter(|(_, recent)| {
-            if query.is_empty() {
-                return true;
-            }
-            recent.project_name.to_lowercase().contains(&query)
-                || recent.project_path.to_lowercase().contains(&query)
-        })
-        .map(|(original_index, recent)| RecentProjectItem {
-            original_index: original_index as i32,
-            title: if recent.project_name.trim().is_empty() {
-                UNTITLED_PROJECT_NAME.into()
-            } else {
-                recent.project_name.clone().into()
-            },
-            subtitle: recent.project_path.clone().into(),
-            is_valid: recent.is_valid,
-            invalid_reason: recent.invalid_reason.clone().unwrap_or_default().into(),
-        })
-        .collect()
-}
-
-pub(crate) fn project_display_name(project: &Project) -> String {
-    project
-        .name
-        .as_ref()
-        .map(|name| name.trim())
-        .filter(|name| !name.is_empty())
-        .map(|name| name.to_string())
-        .unwrap_or_else(|| UNTITLED_PROJECT_NAME.to_string())
-}
-
-pub(crate) fn parse_path_argument(flag: &str) -> Option<PathBuf> {
-    let mut args = env::args().skip(1);
-    while let Some(arg) = args.next() {
-        if arg == flag {
-            return args.next().map(PathBuf::from);
-        }
-    }
-    None
 }
 
 /// Default location for the bundled preset library.
